Notify db committed from tools too
We need this so DB editors can refresh after a tool might write to
them, same as Merger and Importer. Since we cannot control/know what
the tool script does, we need to conservatively assume that it will
write and commit to the DB.
